refactor user dao
