+++ app/src/main/java/com/xepicgamerzx/hotelier/storage/dao/UserDAO.java
@@ -1,18 +1,14 @@
-package com.xepicgamerzx.hotelier.storage.user.data;
+package com.xepicgamerzx.hotelier.storage.dao;
 
 import androidx.room.Dao;
-import androidx.room.Insert;
 import androidx.room.Query;
 
 import com.xepicgamerzx.hotelier.storage.user.model.User;
 
 @Dao
-public interface UserDAO {
-    @Insert
-    void registerUser(User user);
-
+public abstract class UserDAO implements  BaseDao<Void, User>{
     @Query("SELECT * from users where userId=(:userId) and password=(:password)")
-    User login( String userId, String password);
+    public abstract User login( String userId, String password);
 
     // TODO - Add a userFavourites column, so when a user clicks favourite on something, it gets saved.
