+++ app/src/main/java/com/xepicgamerzx/hotelier/user_activities/LoginActivity.java
@@ -1,18 +1,18 @@
 package com.xepicgamerzx.hotelier.user_activities;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.google.android.material.textfield.TextInputEditText;
-import com.xepicgamerzx.hotelier.home_page_activities.MainActivity;
 import com.xepicgamerzx.hotelier.R;
-import com.xepicgamerzx.hotelier.storage.user.data.UserDAO;
-import com.xepicgamerzx.hotelier.storage.user.data.UserDatabase;
+import com.xepicgamerzx.hotelier.home_page_activities.MainActivity;
+import com.xepicgamerzx.hotelier.storage.dao.UserDAO;
+import com.xepicgamerzx.hotelier.storage.hotelier_database.HotelierDatabase;
 import com.xepicgamerzx.hotelier.storage.user.model.User;
 
 public class LoginActivity extends AppCompatActivity {
@@ -39,8 +39,8 @@ public class LoginActivity extends AppCompatActivity {
                     Toast.makeText(getApplicationContext(), "Fill all fields.", Toast.LENGTH_SHORT).show();
                 } else {
                     // Perform query
-                    UserDatabase userDatabase = UserDatabase.getUserDatabase(getApplicationContext());
-                    UserDAO userDao = userDatabase.userDao();
+                    HotelierDatabase hotelierDatabase = HotelierDatabase.getDatabase(getApplicationContext());
+                    UserDAO userDao = hotelierDatabase.userDao();
                     new Thread(new Runnable() {
                         @Override
                         public void run() {
